+++ libs/gltfio/src/FFilamentInstance.h
@@ -117,7 +117,7 @@ struct FFilamentInstance : public FilamentInstance {
     const utils::Entity* getJointsAt(size_t skinIndex) const noexcept;
     void attachSkin(size_t skinIndex, utils::Entity target) noexcept;
     void detachSkin(size_t skinIndex, utils::Entity target) noexcept;
-    const math::mat4f* getInverseBindMatricesAt(size_t skinIndex) const noexcept;
+    math::mat4f const* getInverseBindMatricesAt(size_t skinIndex) const;
 
     void recomputeBoundingBoxes();
 };